Abstracted reused transmuterApr logic into a custom hook
